backend/requirements.txt
fastapi==0.116.1
langchain==0.3.26
langchain_core==0.3.68
langchain_openai==0.3.27
langgraph==0.5.2
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
uvicorn==0.35.0
neo4j==5.15.0
psycopg2-binary==2.9.10
